+++ core/tango-core/src/shadow.rs
@@ -68,7 +68,7 @@ pub struct RoundState {
 }
 
 struct InnerState {
-    match_type: u16,
+    match_type: u8,
     is_offerer: bool,
     round_state: parking_lot::Mutex<RoundState>,
     rng: parking_lot::Mutex<rand_pcg::Mcg128Xsl64>,
@@ -87,7 +87,7 @@ pub struct State(std::sync::Arc<InnerState>);
 
 impl State {
     pub fn new(
-        match_type: u16,
+        match_type: u8,
         is_offerer: bool,
         rng: rand_pcg::Mcg128Xsl64,
         won_last_round: bool,
@@ -105,7 +105,7 @@ impl State {
         }))
     }
 
-    pub fn match_type(&self) -> u16 {
+    pub fn match_type(&self) -> u8 {
         self.0.match_type
     }
 
@@ -156,7 +156,7 @@ impl Shadow {
     pub fn new(
         rom_path: &std::path::Path,
         save_path: &std::path::Path,
-        match_type: u16,
+        match_type: u8,
         is_offerer: bool,
         won_last_round: bool,
         rng: rand_pcg::Mcg128Xsl64,
